Separate PyExpr
