+++ py-polars/src/lazy/mod.rs
@@ -1,7 +1,7 @@
-pub mod dsl;
-
-use dsl::*;
+use polars::lazy::dsl::Operator;
 use polars_lazy::prelude::*;
+
+use crate::PyExpr;
 
 pub(crate) trait ToExprs {
     fn to_exprs(self) -> Vec<Expr>;
@@ -27,4 +27,28 @@ impl ToPyExprs for Vec<Expr> {
         // and has only got one inner field`
         unsafe { std::mem::transmute(self) }
     }
+}
+
+pub fn binary_expr(l: PyExpr, op: u8, r: PyExpr) -> PyExpr {
+    let left = l.inner;
+    let right = r.inner;
+
+    let op = match op {
+        0 => Operator::Eq,
+        1 => Operator::NotEq,
+        2 => Operator::Lt,
+        3 => Operator::LtEq,
+        4 => Operator::Gt,
+        5 => Operator::GtEq,
+        6 => Operator::Plus,
+        7 => Operator::Minus,
+        8 => Operator::Multiply,
+        9 => Operator::Divide,
+        10 => Operator::Modulus,
+        11 => Operator::And,
+        12 => Operator::Or,
+        _ => panic!("not an operator"),
+    };
+
+    polars::lazy::dsl::binary_expr(left, op, right).into()
 }
